requirements.txt
astroid==2.3.1
colorama==0.4.1
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.2
Markdown==3.1.1
mccabe==0.6.1
numpy==1.17.2
openpyxl==3.0.0
pandas==0.25.1
pylint==2.4.2
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
xlrd==1.2.0
